Improve cross-year handling.
I modified _yearLoaded to be a list instead of a single string or None.
We can now have multiple years loaded at once, which is necessary for
expected behavior around the beginning of a new year.  Saving is handled
intelligently so that different years are saved to different files, and
tex output is still provided for just a single year at a time.